Added the basics for a moving platform
And I mean basics. Still need to include adding a end location and the calculation of how much movement to make depending on how much time I want it to take to get there. I did originally implement it, but I had an issue that looking back on I believe was caused by not factoring in DeltaTime. Regardless, it'll be solved tomorrow, 1/14.